+++ PythonClient/airsim/types.py
@@ -37,6 +37,17 @@ class LandedState:
     Landed = 0
     Flying = 1
 
+class WeatherParameter:
+    Rain = 0
+    Roadwetness = 1
+    Snow = 2
+    RoadSnow = 3
+    MapleLeaf = 4
+    RoadLeaf = 5
+    Dust = 6
+    Fog = 7
+    Enabled = 8
+
 class Vector3r(MsgpackMixin):
     x_val = 0.0
     y_val = 0.0